Udpate data code

+++ src/pages/ProyectoDetalle.jsx
@@ -37,6 +37,10 @@ function ProyectoDetalle() {
   };
 
   const handleAsignarUsuarios = async (userId) => {
+    if (authData.user.rol !== 'admin' || authData.user.rol == 'colaborador') {
+      console.error('Solo los administradores pueden asignar usuarios.');
+      return;
+    }
     try {
       const response = await axios.post(`${import.meta.env.VITE_API_URL}/projects/${id}/participants/${userId}`, {}, {
         headers: {
@@ -53,6 +57,10 @@ function ProyectoDetalle() {
   };
 
   const handleDesAsignarUsuarios = async (userId) => {
+    if (authData.user.rol !== 'admin' || authData.user.rol == 'colaborador') {
+      console.error('Solo los administradores pueden desasignar usuarios.');
+      return;
+    }
     try {
       const response = await axios.delete(`${import.meta.env.VITE_API_URL}/projects/${id}/participants/${userId}`, {
         headers: {
@@ -74,100 +82,100 @@ function ProyectoDetalle() {
   }, [proyecto]);
 
   return (
-    <div className="container">
-      {loading && <p>Cargando proyecto...</p>}
-      {error && <p>Error al cargar el proyecto</p>}
+    <div className="container bg-light text-dark p-4 rounded shadow-sm">
+      {loading && <p className="text-muted">Cargando proyecto...</p>}
+      {error && <p className="text-danger">Error al cargar el proyecto</p>}
       {!loading && !error && proyecto && (
         <div>
-          <h1>{proyecto.nombre_proyecto}</h1>
-          <div className="proyecto-descripcion">
-          <div className='clientes'>
-              <button onClick={() => setClientesOpen(!isClientesOpen)} style={{ cursor: 'pointer' }}>
-              <strong>Clientes:</strong>
+          <h1 className="mb-4">{proyecto.nombre_proyecto}</h1>
+          <div className="proyecto-descripcion mb-4">
+            <div className='clientes mb-3'>
+              <button onClick={() => setClientesOpen(!isClientesOpen)} className="btn btn-light card dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded={isClientesOpen}>
+                <strong>Clientes:</strong>
               </button>
-              {isClientesOpen && (
-                <div>
-                  {usuarios.map(usuario => (
-                    usuario.rol === 'cliente' && (
-                      <div key={usuario.id_usuario}>
-                        {usuario.nombre_usuario}
-                      </div>
-                    )
-                  ))}
-                </div>
-              )}
+              <ul className={`dropdown-menu ${isClientesOpen ? 'show' : ''}`}>
+                {usuarios.map(usuario => (
+                  usuario.rol === 'cliente' && (
+                    <li key={usuario.id_usuario} className="dropdown-item text-muted">
+                      {usuario.nombre_usuario}
+                    </li>
+                  )
+                ))}
+              </ul>
             </div>
             <strong>Descripción:</strong>
             {proyecto.descripcion && (
-              <div dangerouslySetInnerHTML={{ __html: decodeHTML(proyecto.descripcion) }} />
+              <div className="mt-2" dangerouslySetInnerHTML={{ __html: decodeHTML(proyecto.descripcion) }} />
             )}
           </div>
-          {(authData.user.rol === 'admin' || authData.user.rol === 'colaborador') && (
+          {(authData.user.rol === 'admin' ) && (
             <>
               <button
                 onClick={() => navigate(`/proyectos/${id}/editar`)}
-                className="button button-edit"
+                className="btn btn-primary me-2"
               >
                 Editar Proyecto
               </button>
-              <button onClick={handleEliminar} className="button button-delete">Eliminar proyecto</button>
+              <button onClick={handleEliminar} className="btn btn-danger">Eliminar proyecto</button>
             </>
           )}
           <button 
             onClick={() => navigate(`/proyectos/${id}/crear-tarea`)}
-            className="button button-create"
+            className="btn btn-success mt-3"
           >
             Crear Tarea
           </button>
-          <div className='container-usuarios'>
-            <div className="usuarios-lista">
-              <button onClick={() => setProjectUsersOpen(!isProjectUsersOpen)} style={{ cursor: 'pointer' }}>
-                Usuarios Asignados
-              </button>
-              <div className="container-asignados">
-                {isProjectUsersOpen && (
-                  <>
-                    {loadingUsuarios && <p>Cargando usuarios...</p>}
-                    {errorUsuarios && <p>Error al cargar los usuarios</p>}
+          {authData.user.rol === 'admin' && (
+            <div className='container-usuarios mt-4'>
+              <div className="usuarios-lista mb-3">
+                <div className="dropdown">
+                  <button onClick={() => setProjectUsersOpen(!isProjectUsersOpen)} className="btn btn-outline-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    Usuarios Asignados
+                  </button>
+                  <ul className={`dropdown-menu ${isProjectUsersOpen ? 'show' : ''}`}>
+                    {loadingUsuarios && <li className="dropdown-item">Cargando usuarios...</li>}
+                    {errorUsuarios && <li className="dropdown-item">Error al cargar los usuarios</li>}
                     {!loadingUsuarios && !errorUsuarios && proyecto.usuarios.map(usuario => (
                       usuario.rol_usuario !== 'admin' && authData.user.id !== usuario.id_usuario ? (
-                        <button
-                          key={usuario.id_usuario}
-                          onClick={() => handleDesAsignarUsuarios(usuario.id_usuario)}
-                          className="button button-user"
-                        >
-                          {authData.user.id === usuario.id_usuario ? 'Yo ' + usuario.nombre_usuario : usuario.nombre_usuario}
-                        </button>
+                        <li key={usuario.id_usuario} className="dropdown-item">
+                          <button
+                            onClick={() => handleDesAsignarUsuarios(usuario.id_usuario)}
+                            className="btn btn-outline-dark"
+                          >
+                            {authData.user.id === usuario.id_usuario ? 'Yo ' + usuario.nombre_usuario : usuario.nombre_usuario}
+                          </button>
+                        </li>
                       ) : null
                     ))}
-                  </>
-                )}
+                  </ul>
+                </div>
+              </div>
+              <div className="usuarios-lista sin-asignar">
+                <div className="dropdown">
+                  <button onClick={() => setOrganizationUsersOpen(!isOrganizationUsersOpen)} className="btn btn-outline-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    Asignar Colaboradores
+                  </button>
+                  <ul className={`dropdown-menu ${isOrganizationUsersOpen ? 'show' : ''}`}>
+                    {loadingUsuarios && <li className="dropdown-item">Cargando usuarios...</li>}
+                    {errorUsuarios && <li className="dropdown-item">Error al cargar los usuarios</li>}
+                    {!loadingUsuarios && !errorUsuarios && usuarios.filter(usuario => 
+                      !proyecto.usuarios.some(asignado => asignado.id_usuario === usuario.id_usuario)
+                    ).map(usuario => (
+                      <li key={usuario.id} className="dropdown-item">
+                        <button
+                          onClick={() => handleAsignarUsuarios(usuario.id_usuario)}
+                          className="btn btn-outline-dark"
+                        >
+                          { authData.user.id === usuario.id_usuario ?  'YO '+ usuario.nombre_usuario  : usuario.nombre_usuario}
+                        </button>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
               </div>
             </div>
-            <div className="usuarios-lista sin-asignar">
-              <button onClick={() => setOrganizationUsersOpen(!isOrganizationUsersOpen)} style={{ cursor: 'pointer' }}>
-                Asignar Colaboradores
-              </button>
-              {isOrganizationUsersOpen && (
-                <>
-                  {loadingUsuarios && <p>Cargando usuarios...</p>}
-                  {errorUsuarios && <p>Error al cargar los usuarios</p>}
-                  {!loadingUsuarios && !errorUsuarios && usuarios.filter(usuario => 
-                    !proyecto.usuarios.some(asignado => asignado.id_usuario === usuario.id_usuario)
-                  ).map(usuario => (
-                    <button
-                      key={usuario.id}
-                      onClick={() => handleAsignarUsuarios(usuario.id_usuario)}
-                      className="button button-user"
-                    >
-                      { authData.user.id === usuario.id_usuario ?  'YO '+ usuario.nombre_usuario  : usuario.nombre_usuario}
-                    </button>
-                  ))}
-                </>
-              )}
-            </div>
-          </div>
-          <p>
+          )}
+          <p className="mt-4">
             <strong>Fecha de fin:</strong>{" "}
             {proyecto.fecha_fin
               ? new Date(proyecto.fecha_fin).toLocaleDateString("es-ES", {
